Windows compiler complaint cleanup
http://codereview.appspot.com/6441055/



git-svn-id: http://skia.googlecode.com/svn/trunk@4796 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ tools/bench_pictures_main.cpp
@@ -239,12 +239,12 @@ static void run_single_benchmark(const SkString& inputPath,
            filename.c_str());
 
     if (options->fTileWidthPercentage > 0) {
-        options->fTileWidth = sk_float_ceil2int(options->fTileWidthPercentage * picture.width()
-                                                / 100);
+        options->fTileWidth = sk_float_ceil2int(
+            float(options->fTileWidthPercentage * picture.width() / 100));
     }
     if (options->fTileHeightPercentage > 0) {
-        options->fTileHeight = sk_float_ceil2int(options->fTileHeightPercentage * picture.height()
-                                                 / 100);
+        options->fTileHeight = sk_float_ceil2int(
+            float(options->fTileHeightPercentage * picture.height() / 100));
     }
 
     options->fBenchmark(&picture, *options);
